+++ Agendamento/templates/index.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html>
+   <head>
+     <Title>Agendmentos</Title>
+   </head>
+   <body>
+Agendamentos
+   </body>
+</html>
