+++ TrilhaDaEvolucao/settings.py
@@ -36,7 +36,8 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'TrilhaDaEvolucao'
+    'TrilhaDaEvolucao',
+    'Agendamento'
 ]
 
 MIDDLEWARE = [
