+++ TrilhaDaEvolucao/templates/NovoAgendamento.html
@@ -1,0 +1,10 @@
+{% extends "base.html" %}
+{% block conteudo %}
+<h2>Novo agendamento</h2>
+<p></p>
+<form method="POST">
+{% csrf_token %}
+{{ form }}
+<button type="submit" class="btn btn-primary">Enviar</button>
+</form>
+{% endblock conteudo %}
