+++ TrilhaDaEvolucao/templates/ajuda.html
@@ -1,0 +1,11 @@
+{% extends "base.html" %}
+
+{% block conteudo %}
+
+<h2>Ajuda</h2>
+
+<h3> Trilha</h3>
+
+<p>da Evolução</p>
+
+{% endblock conteudo %}
