+++ TrilhaDaEvolucao/templates/base.html
@@ -1,0 +1,40 @@
+{% load static %}
+
+<!doctype html>
+
+<html>
+
+<head>
+
+<!-- tags meta obrigatórias -->
+
+<meta charset="utf-8">
+
+<meta name="viewport" content="width=device-width,
+initial-scale=1, shrink-to-fit=no">
+
+<link rel="stylesheet" href="{% static 'css/bootstrap.min.css' %}" >
+
+<title> Trilha</title>
+
+</head>
+
+<body>
+
+{% include 'nav.html' %}
+
+{% block conteudo %}
+
+{% endblock conteudo %}
+
+{% include 'rodape.html' %}
+
+<script src="{% static 'js/jquery.min.js' %}"></script>
+
+<script src="{% static 'js/popper.min.js' %}"></script>
+
+<script src="{% static 'js/bootstrap.min.js' %}"></script>
+
+</body>
+
+</html>
