+++ TrilhaDaEvolucao/templates/index.html
@@ -1,0 +1,10 @@
+{% extends "base.html" %}
+
+{% block conteudo %}
+
+<h2>Trilha da Evolução</h2>
+
+<p>Menu principal</p>
+
+{% endblock conteudo %}
+
