+++ TrilhaDaEvolucao/templates/nav.html
@@ -1,0 +1,72 @@
+<nav class="navbar navbar-expand-lg navbar-light bg-primary">
+
+   <a class="navbar-brand text-light" href="/">
+   
+   Trilha da Evolução
+   
+   </a>
+   
+   <button class="navbar-toggler"
+   
+   type="button"
+   
+   data-toggle="collapse"
+   
+   data-target="#navbarSupportedContent" >
+   
+   <span class="navbar-toggler-icon"></span>
+   
+   </button>
+   
+   <div class="collapse navbar-collapse" id="navbarSupportedContent">
+   
+   <ul class="navbar-nav mr-auto">
+   
+   <li class="nav-item">
+   
+   <a class="nav-link text-light" href="/">
+   
+   Página Principal
+   
+   </a>
+   
+   </li>
+   <li class="nav-item">
+   
+      <a class="nav-link text-light" href="/">
+      
+      Agendamentos
+      
+      </a>
+      
+      </li>
+      <li class="nav-item">
+   
+         <a class="nav-link text-light" href="/">
+         
+         Evolução
+         
+         </a>
+         
+         </li>
+
+            <li class="nav-item">
+   
+   <a class="nav-link text-light" href="/admin">
+   
+   Cadastros
+   
+   </a>
+   
+   </li>
+   <li class="nav-item">
+   
+   <a class="nav-link text-light" href="/ajuda/">Ajuda</a>
+   
+   </li>
+   
+   </ul>
+   
+   </div>
+   
+   </nav>
