+++ TrilhaDaEvolucao/templates/rodape.html
@@ -1,0 +1,6 @@
+
+<div class="navbar fixed-bottom bg-info">
+
+   Projeto Integrador - 2021 - Univesp - G002 - Polo de São José dos Campos - V 0.0.1
+   
+   </div>
